+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### 🛑 Breaking changes 🛑
 
+- Remove `configunmarshaler.Unmarshaler` interface, per deprecation comment )#5348)
 - Remove deprecated pdata funcs/structs from v0.50.0 (#5345)
 
 ### 🚩 Deprecations 🚩